requirements.txt
PyGithub
ipython
pyyaml